remove hashicorp code
