+++ weed/go.mod
@@ -3,14 +3,11 @@ module weed
 go 1.17
 
 require (
-	github.com/Jille/raft-grpc-transport v1.4.0
 	github.com/boltdb/bolt v1.3.1
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/golang/protobuf v1.5.2
 	github.com/google/btree v1.1.2
 	github.com/gorilla/mux v1.8.0
-	github.com/hashicorp/raft v1.5.0
-	github.com/hashicorp/raft-boltdb/v2 v2.2.2
 	github.com/klauspost/crc32 v1.2.0
 	github.com/seaweedfs/raft v1.1.2
 	github.com/soheilhy/cmux v0.1.5
@@ -20,21 +17,11 @@ require (
 )
 
 require (
-	github.com/armon/go-metrics v0.4.1 // indirect
-	github.com/fatih/color v1.13.0 // indirect
 	github.com/golang/snappy v0.0.0-20180518054509-2e65f85255db // indirect
-	github.com/hashicorp/errwrap v1.0.0 // indirect
-	github.com/hashicorp/go-hclog v1.5.0 // indirect
-	github.com/hashicorp/go-immutable-radix v1.3.1 // indirect
-	github.com/hashicorp/go-msgpack v1.1.5 // indirect
-	github.com/hashicorp/go-multierror v1.1.1 // indirect
-	github.com/hashicorp/golang-lru v0.5.4 // indirect
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	go.etcd.io/bbolt v1.3.5 // indirect
+	github.com/stretchr/testify v1.8.2 // indirect
 	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	google.golang.org/genproto v0.0.0-20210903162649-d08c68adba83 // indirect
 	google.golang.org/protobuf v1.27.1 // indirect
+	gopkg.in/yaml.v2 v2.2.5 // indirect
 )
