+++ weed/server/cluster_commands.go
@@ -1,9 +1,6 @@
 package weed_server
 
 import (
-	"encoding/json"
-	"fmt"
-	hashicorpRaft "github.com/hashicorp/raft"
 	"weed/glog"
 	"weed/storage"
 	"weed/topology"
@@ -33,20 +30,4 @@ func (c *MaxVolumeIdCommand) Apply(server raft.Server) (interface{}, error) {
 	glog.V(0).Infoln("max volume id", before, "==>", topo.GetMaxVolumeId())
 
 	return nil, nil
-}
-
-func (s *MaxVolumeIdCommand) Persist(sink hashicorpRaft.SnapshotSink) error {
-	b, err := json.Marshal(s)
-	if err != nil {
-		return fmt.Errorf("marshal: %v", err)
-	}
-	_, err = sink.Write(b)
-	if err != nil {
-		sink.Cancel()
-		return fmt.Errorf("sink.Write(): %v", err)
-	}
-	return sink.Close()
-}
-
-func (s *MaxVolumeIdCommand) Release() {
 }
